fix: Car Util 일부 이동
